Client profiles are deleted
For now, it assumes that there will always be a "full" profile packaged
with any "client" profile. This may need to be updated in the future if
that isn't always true.

See profiles here:

https://docs.nuget.org/ndocs/create-packages/supporting-multiple-target-frameworks

Also changed the 3.0 and 3.5 to report correctly even if a profile tag
is in the folder name.