Added Unity 5.2 Packager project (#130)
This is a Unity 5.2 project that is used to package up NuGetForUnity into a .unitypackage file to be uploaded to the Unity Asset Store.

This project is basically used to house the Unity 5.2 .meta files to ensure that the GUIDs stay the same for the assets. If the project and .meta files are automatically generated by the build server, the GUIDs would be different and all files would be seen as changed for users updating.