Deleting unnecessary folders after update packages.json from git.
Description of the problem that is solved:
When one developer uninstall any package and commit it. The folder remove on this developer pc, but other developer pull this changes, Assets/Packages folder in ignore, open Unity and automatically start restore packages. Folder still exist because there is no code for check this.